% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_grade.R
\name{ae_plot_grade}
\alias{ae_plot_grade}
\title{Graphic representation of AEs}
\usage{
ae_plot_grade(
  df_ae,
  ...,
  df_enrol,
  type = c("max", "sup", "eq"),
  type2 = c("relative", "absolute"),
  arm = NULL,
  grade = "AEGR",
  subjid = "SUBJID"
)
}
\arguments{
\item{df_ae}{adverse event dataset, one row per AE, containing subjid, soc, and grade}

\item{...}{unused}

\item{df_enrol}{enrollment dataset, one row per patient, containing subjid (and arm if needed). All patients should be in this dataset.}

\item{type}{default to \code{c("max", "sup", "eq")}. \code{max} computes the maximum AE grade per patient, \code{sup} computes the number of patients having experienced at least one AE of grade higher than X, and \code{eq} computes the number of patients having experienced at least one AE of grade equal to X.}

\item{type2}{whether to present patients as proportions (\code{relative}) or as counts (\code{absolute})}

\item{arm}{name of the treatment column in \code{df_enrol}. Case-insensitive. Can be set to \code{NULL} to not group.}

\item{grade}{name of the AE grade column in \code{df_ae}. Case-insensitive.}

\item{subjid}{name of the patient ID in both \code{df_ae} and \code{df_enrol}. Case-insensitive.}
}
\value{
a ggplot
}
\description{
Produce a graphic representation of AE, counting AE as bars for each patient, colored by grade. Can be faceted by treatment arm.
}
