% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_grade.R
\name{ae_plot_grade_n}
\alias{ae_plot_grade_n}
\title{Graphic representation of AEs}
\usage{
ae_plot_grade_n(
  df_ae,
  df_enrol,
  low = "#ffc425",
  high = "#d11141",
  arm = "ARM",
  grade = "AEGR",
  subjid = "SUBJID"
)
}
\arguments{
\item{df_ae}{adverse event table, one row per AE, containing subjid, soc, and grade}

\item{df_enrol}{enrollment table, one row per patient, containing subjid (and arm if needed)}

\item{low}{the color of Grade 1 AE}

\item{high}{the color of Grade 5 AE}

\item{arm}{name of the treatment column in \code{df_enrol}. Case-insensitive. Can be set to \code{NULL} to not group.}

\item{grade}{name of the AE grade column in \code{df_ae}. Case-insensitive.}

\item{subjid}{name of the patient ID in both \code{df_ae} and \code{df_enrol}. Case-insensitive.}
}
\value{
a ggplot
}
\description{
Produce a graphic representation of AE, counting AE as bars for each patient, colored by grade. Can be faceted by treatment arm.
}
\examples{
tm = edc_example_ae()
ae_plot_grade_n(df_ae=tm$ae, df_enrol=tm$enrolres)
ae_plot_grade_n(df_ae=tm$ae, df_enrol=tm$enrolres, arm=NULL)
}
\seealso{
\code{\link[=ae_table_grade_max]{ae_table_grade_max()}}, \code{\link[=ae_table_grade_n]{ae_table_grade_n()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade_max]{ae_plot_grade_max()}}, \code{\link[=ae_plot_grade_n]{ae_plot_grade_n()}}
}
