% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_soc.R
\name{ae_plot_soc}
\alias{ae_plot_soc}
\title{Graphic representation of AEs by soc (Butterfly plot)}
\usage{
ae_plot_soc(
  df_ae,
  ...,
  df_enrol,
  arm = "ARM",
  subjid = "SUBJID",
  soc = "AESOC",
  severe = "SEVERE"
)
}
\arguments{
\item{df_ae}{adverse event table, one row per AE, containing subjid, soc, and grade}

\item{...}{unused}

\item{df_enrol}{enrollment table, one row per patient, containing subjid (and arm if needed). All patients should be in this table.}

\item{arm}{name of the treatment column in \code{df_enrol}. Case-insensitive. Can be set to \code{NULL} to not group.}

\item{subjid}{name of the patient ID in both \code{df_ae} and \code{df_enrol}. Case-insensitive.}

\item{soc}{name of the SOC column in \code{df_ae}. Case-insensitive. Grade will be considered 0 if missing(e.g. if patient if absent from \code{df_ae}).}

\item{severe}{name of the logical column in \code{df_ae} telling wheter an AE is severe. Case-insensitive.}
}
\value{
a crosstable (dataframe)
}
\description{
Produces a graphic representation of AE, counting the maximum grade each patient experienced, colored by treatment arm. Returns up to 3 representations if \code{arm!=NULL}.
}
\details{
The function \code{ae_plot_soc()} creates a summary table of the maximum AE grade experienced per each patient.
The resulting crosstable can be piped to \code{as_flextable()} to get a nicely formatted flextable.
}
\examples{

tm = edc_example_ae()
tm$ae \%>\% 
  #dplyr::mutate(severe = aeser=="Yes") \%>\% 
  dplyr::mutate(severe = aegr>=3) \%>\% 
  ae_plot_soc(df_enrol=tm$enrolres)
}
\seealso{
\code{\link[=ae_table_grade_max]{ae_table_grade_max()}}, \code{\link[=ae_table_grade_n]{ae_table_grade_n()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade_max]{ae_plot_grade_max()}}, \code{\link[=ae_plot_grade_n]{ae_plot_grade_n()}}
}
