% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_grade.R
\name{ae_table_grade_n}
\alias{ae_table_grade_n}
\title{Title}
\usage{
ae_table_grade_n(
  df_ae,
  df_enrol,
  arm = "ARM",
  grade = "AEGR",
  subjid = "SUBJID",
  soc = "AESOC",
  total = FALSE,
  digits = 0
)
}
\arguments{
\item{df_ae}{adverse event table, one row per AE, containing subjid, soc, and grade}

\item{df_enrol}{enrollment table, one row per patient, containing subjid (and arm if needed)}

\item{arm}{name of the treatment column in \code{df}. Case-insensitive. Can be set to \code{NULL} to not group.}

\item{subjid, grade}{names of the other relevant columns in \code{df}. Case-insensitive.}

\item{soc}{name of the SOC column in \code{df}. Case-insensitive. Grade will be considered 0 is missing.}

\item{total}{whether to add totals}

\item{digits}{number of signigicant digits for percentages}
}
\value{
a crosstable
}
\description{
Title
}
\examples{
\dontrun{
tm = edc_example_ae()

ae_table_grade_n(df_ae=tm$ae, df_enrol=tm$enrolres) \%>\% 
  as_flextable() \%>\% 
  flextable::add_footer_lines("Percentages are given as the proportion of patients presenting at least one AE of given grade")

ae_table_grade_n(df_ae=tm$ae, df_enrol=tm$enrolres, arm=NULL) \%>\% 
  as_flextable(by_header=F) \%>\% 
  flextable::set_header_labels(values=c("","","N (\%)")) 
  
#To get SAE only, filter df_ae first
tm$ae \%>\% filter(sae==TRUE) \%>\% ae_table_grade_n(df_enrol=tm$enrolres, arm=NULL)
}
}
