% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_soc.R
\name{ae_table_soc_legacy}
\alias{ae_table_soc_legacy}
\title{for each patient/soc, detect if each grade satisfies the specified
condition (max/eq/sup)}
\usage{
ae_table_soc_legacy(
  df_ae,
  ...,
  df_enrol,
  variant = c("max", "sup", "eq"),
  arm = "ARM",
  term = "AETERM",
  soc = "AESOC",
  grade = "AEGR",
  subjid = "SUBJID",
  sort_by_ae = TRUE,
  total = TRUE,
  digits = 0,
  warn_miss = FALSE
)
}
\description{
for each patient/soc, detect if each grade satisfies the specified
condition (max/eq/sup)
}
