% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assert_no_rows}
\alias{assert_no_rows}
\title{Assert that a dataframe has no rows}
\usage{
assert_no_rows(df, msg = NULL)
}
\arguments{
\item{df}{a dataframe}

\item{msg}{(optional) a custom message to be output (e.g. with the underlying reason)}
}
\value{
nothing
}
\description{
Check that a dataframe is empty in a pipeable style. \cr
Mostly useful for sanity checks.
}
\examples{
tm = edc_example()
tm$db0 \%>\% dplyr::filter(age>100) \%>\% assert_no_rows()
}
