% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{build_lookup}
\alias{build_lookup}
\title{Generate a lookup table}
\usage{
build_lookup(data_list)
}
\arguments{
\item{data_list}{a list containing at least 1 dataframe}
}
\value{
a dataframe summarizing column names and labels
}
\description{
Generate a lookup table
}
\examples{
x = edc_example()
x$.lookup=NULL
lk = build_lookup(x)
lk
lk \%>\% tidyr::unnest(c(names, labels))  

}
\seealso{
\code{\link[=extend_lookup]{extend_lookup()}}, \code{\link[=edc_lookup]{edc_lookup()}}
}
