% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_subjid}
\alias{check_subjid}
\title{Check the completion of the subject ID column}
\usage{
check_subjid(x, ref = getOption("edc_subjid_ref"))
}
\arguments{
\item{x}{the subject ID column to check, or a dataframe which ID column will be guessed}

\item{ref}{the reference for subject ID. Should usually be set through \code{edc_options(edc_subjid_ref=xxx)}. See example.}
}
\value{
nothing, called for errors/warnings
}
\description{
Compare a subject ID vector to the study's reference subject ID (usually something like \code{enrolres$subjid}).
}
\examples{
tm = edc_example()
load_list(tm)
options(edc_subjid_ref=db0$SUBJID)
#usually, you set something like:
#options(edc_subjid_ref=enrolres$subjid)
check_subjid(db1)
db1 \%>\% dplyr::filter(SUBJID>1) \%>\% check_subjid()
check_subjid(c(db1$SUBJID, 99, 999))
}
