% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_example}
\alias{data_example}
\alias{edc_example_mixed}
\alias{edc_example}
\alias{edc_example_plot}
\alias{edc_example_ae}
\title{Example databases}
\usage{
edc_example_mixed(N = 100, seed = 42)

edc_example(N = 50, seed = 42, outdated = FALSE)

edc_example_ae(N = 50, seed = 42)
}
\arguments{
\item{N}{the number of patients}

\item{seed}{the random seed}

\item{outdated}{simulate times after the data extraction date}
}
\value{
a list of tables
}
\description{
List of tables used in EDCimport examples:
\itemize{
\item \code{edc_example()} can be used as the result of reading functions (such as \code{\link[=read_trialmaster]{read_trialmaster()}} or \code{\link[=read_all_sas]{read_all_sas()}}) and to test \code{\link[=edc_swimmerplot]{edc_swimmerplot()}}
\item \code{edc_example_mixed()} can be used to test \code{\link[=split_mixed_datasets]{split_mixed_datasets()}}
}
}
