% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{edc_clean_names}
\alias{edc_clean_names}
\title{Clean up the names of all datasets}
\usage{
edc_clean_names(database)
}
\arguments{
\item{database}{an \link{edc_database} object, from \code{\link[=read_trialmaster]{read_trialmaster()}} or other EDCimport reading functions.}
}
\value{
an \link{edc_database} object
}
\description{
Clean the names of all the datasets in the database, so that columns are only lowercase letters,
numbers, and \verb{_}.
}
\examples{
#db = read_trialmaster("filename.zip", pw="xx")
db = edc_example() \%>\% 
  edc_clean_names()
names(db$enrol)
}
