% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read.R
\docType{class}
\name{edc_database}
\alias{edc_database}
\title{EDCimport Database}
\description{
This class of object represents a database, as the result of an EDCimport reading function.
It has its own \code{print()} method.
}
\section{Functions returning \code{edc_database} objects}{

As per now, reading functions are: \code{\link[=read_trialmaster]{read_trialmaster()}}, \code{\link[=read_all_sas]{read_all_sas()}}, \code{\link[=read_all_xpt]{read_all_xpt()}}, and \code{\link[=read_all_csv]{read_all_csv()}}.
}

\section{Structure}{

While it is not usually useful to query them, an \code{edc_database} object is a named list containing:
\itemize{
\item all the datasets from the source files
\item \code{datetime_extraction} and \code{date_extraction} the inferred date of data extraction
\item \code{.lookup} a temporary copy of the lookup table
}
}

\seealso{
\code{\link[=read_trialmaster]{read_trialmaster()}}
}
