% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{edc_example}
\alias{edc_example}
\alias{edc_example_plot}
\title{Example database}
\usage{
edc_example(N = 50, seed = 42, outdated = FALSE)
}
\arguments{
\item{N}{the number of patients}

\item{seed}{the random seed}

\item{outdated}{whether to simulate times after the data extraction date}
}
\value{
A list of tables of class \code{edc_database}.
}
\description{
A list of tables that simulates the extraction of a clinical database. Used in \code{EDCimport} examples and tests.
}
