% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{edc_example_multiple}
\alias{edc_example_multiple}
\title{Example for \code{compare_databases()}}
\usage{
edc_example_multiple()
}
\description{
This functions returns a list of 3 \code{edc_example()} instances, slightly different from one another.
}
\keyword{internal}
