% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{edc_inform_code}
\alias{edc_inform_code}
\title{Shows how many code you wrote}
\usage{
edc_inform_code(main = "main.R", Rdir = "R/")
}
\arguments{
\item{main}{the main R file, which sources the other ones}

\item{Rdir}{the R directory, where sourced R files are located}
}
\value{
Nothing
}
\description{
Shows how many code you wrote
}
