% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{edc_lookup}
\alias{edc_lookup}
\alias{get_lookup}
\title{Retrieve the lookup table from options}
\usage{
edc_lookup(..., check = TRUE)
}
\arguments{
\item{...}{passed on to \code{\link[dplyr:arrange]{dplyr::arrange()}}}

\item{check}{whether to check for internal consistency}
}
\value{
the lookup dataframe summarizing the database import
}
\description{
Retrieve the lookup table from options
}
\examples{
db = edc_example()
load_database(db)
edc_lookup()
edc_lookup(dataset)
}
