% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{edc_peek_options}
\alias{edc_peek_options}
\title{See which \code{EDCimport} option is currently set.}
\usage{
edc_peek_options(keep_null = FALSE)
}
\arguments{
\item{keep_null}{set to TRUE to get a list}
}
\value{
A named list of EDCimport options
}
\description{
See which \code{EDCimport} option is currently set.
}
