% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{edc_reset_options}
\alias{edc_reset_options}
\title{Reset all \code{EDCimport} options}
\usage{
edc_reset_options(
  except = c("edc_lookup", "trialmaster_pw", "path_7zip"),
  quiet = FALSE
)
}
\arguments{
\item{except}{options that are not reset by default}

\item{quiet}{set to \code{TRUE} to remove the message.}
}
\value{
Nothing, called for its side effects
}
\description{
Reset all \code{EDCimport} options
}
