% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_excel.R
\name{edc_save_db_to_excel}
\alias{edc_save_db_to_excel}
\title{Save the database as an Excel file}
\usage{
edc_save_db_to_excel(
  ...,
  datasets = get_datasets(),
  filename = NULL,
  open = TRUE
)
}
\arguments{
\item{...}{unused}

\item{datasets}{a named list of dataframes. Default to the TM export.}

\item{filename}{the path to the Excel output file. Default to a dated file in the "data" folder.}

\item{open}{whether to open the Excel file afterward. Default to \code{TRUE}.}
}
\value{
nothing
}
\description{
Because RStudio is not very good at showing data, it can be more convenient to browse the
database using MS Excel. This function turns the whole TM export (or any named list of datasets)
into an Excel workbook, with one tab for each dataset.
}
\examples{
\dontrun{
  tm = edc_example()
  load_list(tm)  
  edc_save_db_to_excel() #default arguments are usually OK
}
}
