% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{edc_unify_subjid}
\alias{edc_unify_subjid}
\alias{harmonize_subjid}
\title{Harmonize the subject ID of the database}
\usage{
edc_unify_subjid(database, preprocess = NULL, col_subjid = NULL)
}
\arguments{
\item{database}{an \link{edc_database} object, from \code{\link[=read_trialmaster]{read_trialmaster()}} or other EDCimport reading functions.}

\item{preprocess}{an optional function to modify the subject ID column. Default to \code{as.numeric()} if applicable and \code{identity()} otherwise. See examples.}

\item{col_subjid}{the names of the columns holding the subject ID (as character)}
}
\value{
database, with subject id modified
}
\description{
Turns the subject ID columns of all datasets into a factor containing levels for all
the subjects of the database. Avoid problems when joining tables, and some checks can
be performed on the levels.
}
\examples{
db = edc_example()
db$enrol = head(db$enrol, 10)
db$enrol$subjid \%>\% head()
db = edc_unify_subjid(db)
db$enrol$subjid \%>\% head()
db = edc_unify_subjid(db, preprocess=function(x) paste0("#", x))
db$enrol$subjid \%>\% head()
}
