% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{edc_warn_extraction_date}
\alias{edc_warn_extraction_date}
\title{Warn if extraction is too old}
\usage{
edc_warn_extraction_date(max_days = 30)
}
\arguments{
\item{max_days}{the max acceptable age of the data}
}
\value{
nothing
}
\description{
Warn if extraction is too old
}
\examples{
tm = edc_example()
load_list(tm)
edc_warn_extraction_date()
}
