% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{EDCimport_options}
\alias{EDCimport_options}
\title{Options used in \code{EDCimport}}
\description{
Here are all the options used in \code{EDCimport}:
\itemize{
\item \code{trialmaster_pw}: the password for the archive read by \code{\link[=read_trialmaster]{read_trialmaster()}}
\item \code{edc_lookup}: the default data used by \code{\link[=find_keyword]{find_keyword()}}
\item \code{path_7zip}: the path to 7-zip directory if not in the PATH, usually "C:/Program Files/7-Zip/" on Windows. See \code{\link[=extract_7z]{extract_7z()}} for more information.
}
}
