% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_soc.R
\name{evaluate_grades}
\alias{evaluate_grades}
\title{for each patient/soc, detect if each grade satisfies the specified
condition (max/eq/sup)}
\usage{
evaluate_grades(gr, variant)
}
\description{
for each patient/soc, detect if each grade satisfies the specified
condition (max/eq/sup)
}
