% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extend_lookup}
\alias{extend_lookup}
\title{TODO extend_lookup}
\usage{
extend_lookup(
  id = getOption("edc_id", "SUBJID"),
  lookup = getOption("edc_lookup", NULL)
)
}
\arguments{
\item{id}{}

\item{lookup}{}
}
\description{
TODO extend_lookup
}
