% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7zip.R
\name{get_7z_dir}
\alias{get_7z_dir}
\title{Retrieve the path to 7zip executable directory}
\usage{
get_7z_dir()
}
\value{
the path to 7zip executable directory
}
\description{
Retrieve the path to 7zip executable directory. Use \code{options(path_7zip="path/to/7zip")} to change its behavior.
}
