% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_datasets}
\alias{get_datasets}
\title{Retrieve the datasets as a list of data.frames}
\usage{
get_datasets(lookup = edc_lookup(), envir = parent.frame())
}
\arguments{
\item{lookup}{the lookup table}

\item{envir}{(internal use)}
}
\value{
a list of all datasets
}
\description{
Get the datasets from the lookup table as a list of data.frames.
}
