% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_key_cols}
\alias{get_key_cols}
\title{Important column names}
\usage{
get_key_cols(lookup = getOption("edc_lookup", NULL))
}
\arguments{
\item{lookup}{the lookup table}
}
\value{
a list of characters
}
\description{
Retreive names of \code{patient_id} (usually "SUBJID" and "PATNO") and \code{crfname} (usually "CRFNAME") from the actual names of the datasets
}
