% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lastnews_table}
\alias{lastnews_table}
\title{Get a table with the latest date for each patient}
\usage{
lastnews_table(except = NULL, with_ties = FALSE)
}
\arguments{
\item{except}{the datasets that should not be searched}

\item{with_ties}{in case of tie, whether to return the first \code{origin} (FALSE) or all the origins that share this tie (TRUE).}
}
\value{
a dataframe
}
\description{
This function search for date columns in every tables and returns the latest date
for each patient with the variable it comes from.
}
\examples{
tm = edc_example_plot()
load_list(tm)
lastnews_table()
lastnews_table(except="db3")
lastnews_table(except="db3$date9")
}
