% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialmaster.R
\name{load_data}
\alias{load_data}
\title{Load the data}
\usage{
load_data(
  target,
  update = TRUE,
  input_folder = "data/",
  return_datetime = FALSE
)
}
\arguments{
\item{target}{the name of the RData to load, or "last" to load the most recent one}

\item{update}{whether to unzip some new archives first}

\item{input_folder}{default to "data/"}

\item{return_datetime}{returns a datetime if \code{TRUE}, a date if \code{FALSE}}
}
\value{
nothing, called for its side-effect
}
\description{
Load the data
}
