% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_database}
\alias{load_database}
\alias{load_list}
\title{Load a list in an environment}
\usage{
load_database(db, env = parent.frame(), remove = TRUE)
}
\arguments{
\item{db}{an \link{edc_database} object (to be fair, any list would do)}

\item{env}{the environment onto which the list should be loaded}

\item{remove}{if \code{TRUE}, \code{db} will be removed from the environment afterward}
}
\value{
nothing, called for its side-effect
}
\description{
Load a list in an environment
}
\examples{
db = edc_example()
load_database(db, remove=FALSE)
print(db)
print(lengths(db))
}
