% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_list}
\alias{load_list}
\title{Load a list in an environment}
\usage{
load_list(x, env = parent.frame(), remove = TRUE)
}
\arguments{
\item{x}{a list}

\item{env}{the environment onto which the list should be loaded}

\item{remove}{if \code{TRUE}, \code{x} will be removed from the environment afterward}
}
\value{
nothing, called for its side-effect
}
\description{
Load a list in an environment
}
\examples{

x=list(a=1, b=mtcars)
load_list(x, remove=FALSE)
print(a)
print(nrow(b))

}
