% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_csv.R
\name{read_all_csv}
\alias{read_all_csv}
\title{Title}
\usage{
read_all_csv(
  path,
  ...,
  label_dict = NULL,
  clean_names_fun = NULL,
  datetime_extraction = "guess",
  verbose = getOption("edc_read_verbose", 1)
)
}
\arguments{
\item{path}{}

\item{...}{}

\item{label_dict}{}

\item{clean_names_fun}{}

\item{datetime_extraction}{}

\item{verbose}{}
}
\description{
Title
}
