% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_xpt.R
\name{read_all_xpt}
\alias{read_all_xpt}
\alias{read_tm_all_xpt}
\title{Read all \code{.xpt} files in a directory}
\usage{
read_all_xpt(
  path,
  ...,
  format_file = "procformat.sas",
  datetime_extraction = "guess",
  subdirectories = FALSE,
  verbose = getOption("edc_read_verbose", 1),
  clean_names_fun = NULL,
  directory = "deprecated",
  key_columns = "deprecated"
)
}
\arguments{
\item{path}{[\code{character(1)}]\cr the path to the directory containing all \code{.xpt} files.}

\item{...}{unused}

\item{format_file}{[\code{character(1)}]\cr the path to the file that should be used to apply formats. See section "Format file" below. Use \code{NULL} to not apply formats.}

\item{datetime_extraction}{[\code{POSIXt(1)}]\cr the datetime of the data extraction. Default to the most common date of last modification in \code{path}.}

\item{subdirectories}{[\code{logical(1)}]\cr whether to read subdirectories}

\item{verbose}{[\code{numeric(1)}]\cr one of \code{c(0, 1, 2)}. The higher, the more information will be printed.}

\item{clean_names_fun}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} use \code{\link[=edc_clean_names]{edc_clean_names()}} instead.}

\item{directory}{deprecated in favour for \code{path}}

\item{key_columns}{deprecated}
}
\value{
a list containing one dataframe for each \code{.xpt} file in the folder, the extraction date (\code{datetime_extraction}), and a summary of all imported tables (\code{.lookup}).
}
\description{
Read all \code{.xpt} files in a directory (unzipped TrialMaster archive). \cr
If \verb{7zip} is installed, you should probably rather use \code{\link[=read_trialmaster]{read_trialmaster()}} instead. \cr
If a \code{procformat.sas} file exists in the directory, formats will be applied.
}
\section{Format file}{

\code{format_file} should contain the information about SAS formats. It can be either:
\itemize{
\item a \code{procformat.sas} file, containing the whole PROC FORMAT
\item or a data file (.csv or .sas7bdat) containing 3 columns:
\itemize{
\item \code{FMTNAME} the SAS format name (repeated)
\item \code{START} the variable level
\item \code{LABEL} the label associated to the level
}

You can get this datafile \href{https://blogs.sas.com/content/sgf/2017/12/04/controlling-your-formats/}{from SAS} using \verb{PROC FORMAT} with option \code{CNTLOUT}.
Otherwise, you can use \code{options(edc_var_format_name="xxx", edc_var_level="xxx", edc_var_label="xxx")} to specify different column names.
}
}

\seealso{
Other EDCimport reading functions: 
\code{\link{read_all_csv}()},
\code{\link{read_all_sas}()},
\code{\link{read_trialmaster}()}
}
\concept{EDCimport reading functions}
