% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_list}
\alias{save_list}
\title{Save a list as \code{.RData} file}
\usage{
save_list(x, filename)
}
\arguments{
\item{x}{a list}

\item{filename}{the filename, with the \code{.RData} extension.}
}
\value{
nothing, called for its side-effect
}
\description{
Save a list as \code{.RData} file
}
\examples{
x=list(a=1, b=mtcars)
save_list(x, "test.RData")
load("test.RData")
file.remove("test.RData")
print(a)
print(nrow(b))
}
