% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_sessioninfo}
\alias{save_sessioninfo}
\title{Save \code{sessionInfo()} output}
\usage{
save_sessioninfo(path = "check/session_info.txt", with_date = TRUE)
}
\arguments{
\item{path}{target path to write the file}

\item{with_date}{whether to insert the date before the file extension}
}
\value{
nothing
}
\description{
Save \code{sessionInfo()} output into a text file.
}
\examples{
\dontrun{
   save_sessioninfo()
}
}
