% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_project_name}
\alias{set_project_name}
\title{Set the project name}
\usage{
set_project_name(db, name)
}
\arguments{
\item{db}{the \link{edc_database}}

\item{name}{the project name}
}
\value{
nothing
}
\description{
Set or override the project name
}
\examples{
db = edc_example() \%>\% 
 set_project_name("My great project")
edc_lookup()
}
