% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_mixed.R
\name{split_mixed_datasets}
\alias{split_mixed_datasets}
\title{Split mixed datasets}
\usage{
split_mixed_datasets(
  id,
  ...,
  verbose = TRUE,
  output_code = FALSE,
  lookup = getOption("edc_lookup", NULL)
)
}
\arguments{
\item{id}{the patient identifier, probably "SUBJID". Should be shared by all datasets.}

\item{...}{not used}

\item{verbose}{whether to print informations about the process.}

\item{output_code}{whether to print the code. Can also be a file path.}

\item{lookup}{the lookup table}
}
\value{
a list of the new long and short tables. Use \code{\link[=load_list]{load_list()}} to load them in the global environment.
}
\description{
Split mixed tables, i.e. tables that hold both long data (N values per patient) and short data (one value per patient, duplicated on N lines), into one long table and one short table.
}
\examples{
\dontrun{
tm = read_trialmaster("filename.zip", pw="xx")
load_list(tm)

mixed = split_mixed_datasets("SUBJID")
load_list(mixed) 
}
}
