% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimmerplot.R
\name{swimmerplot}
\alias{swimmerplot}
\title{Swimmer plot of all dates columns}
\usage{
swimmerplot(
  .lookup = getOption("edc_lookup", NULL),
  ...,
  id = "SUBJID",
  group = NULL,
  origin = NULL
  time_unit = c("days", "weeks", "months", "years"),
  aes_color = c("variable", "label"),
  plotly = TRUE
)
}
\arguments{
\item{.lookup}{the lookup table, loaded along with the database or result of \code{\link[=get_lookup]{get_lookup()}}}

\item{...}{not used}

\item{id}{the patient identifier}

\item{origin}{a variable to consider as time 0, given as "dataset$column"}

\item{group}{a grouping variable, given as "dataset$column"}

\item{origin}{a variable to consider as time 0, given as "dataset$column"}

\item{time_unit}{if \code{origin!=NULL}, the unit to measure time. One of \code{c("days", "weeks", "months", "years")}.}

\item{aes_color}{either \code{variable} ("{dataset} - {column}") or \code{label} (the column label)}

\item{plotly}{whether to use \code{{plotly}} to get an interactive plot}
}
\value{
either a \code{plotly} or a \code{ggplot}
}
\description{
Join all tables from \code{.lookup$dataset} on \code{id}
}
\examples{
#tm = read_trialmaster("filename.zip", pw="xx")
tm = edc_example_plot()
load_list(tm)
p = swimmerplot(.lookup)
p2 = swimmerplot(.lookup, origin="db0$date_naissance", time_unit="weeks")
p3 = swimmerplot(.lookup, group="db0$group", aes_color="label")
\dontrun{
#save the plotly plot as HTML to share it
htmlwidgets::savewidget(p, "swimmerplot.html", selfcontained=TRUE)
}
}
\examples{
\dontrun{
  p = swimmerplot()
  htmlwidgets::savewidget(p, "swimmerplot.html", selfcontained=TRUE)
}
}
