% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_mixed.R
\name{unify}
\alias{unify}
\title{Unify a vector}
\usage{
unify(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector of length 1
}
\description{
Turn a vector of length N to a vector of length 1 after checking that there is only one unique value. Useful to safely flatten a duplicated table. This preserves the \code{label} attribute if set.
}
\examples{
unify(c(1,1,1,1))
#unify(c(1,1,2,1)) #warning
}
