% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialmaster.R
\name{update_data_folder}
\alias{update_data_folder}
\title{Update the "data" folder}
\usage{
update_data_folder(
  input_folder = "data/",
  output_file = "data/atezo_{date}.RData",
  path_7zip = "C:\\\\Program Files\\\\7-Zip\\\\",
  verbose = TRUE
)
}
\arguments{
\item{input_folder}{the input folder, containing the archives}

\item{output_file}{glue pattern for the .RData destination. Variable {date} can be used.}

\item{path_7zip}{path to the 7-zip software}

\item{verbose}{the verbosity level}
}
\value{
nothing, called for its side-effect
}
\description{
Update the "data" folder
}
