% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{waterfall_plot}
\alias{waterfall_plot}
\title{Generate a waterfall plot}
\usage{
waterfall_plot(
  data_recist,
  rc_date = "RCDT",
  rc_sum = "RCTLSUM",
  rc_resp = "RCRESP",
  rc_new = "RCNEW",
  rc_new_needle = "yes",
  arm = NULL,
  warn_missing = TRUE
)
}
\arguments{
\item{data_recist}{}

\item{rc_date}{}

\item{rc_sum}{}

\item{rc_resp}{}

\item{arm}{}

\item{warn_missing}{}
}
\value{
a ggplot
}
\description{
Generate a waterfall plot
}
\examples{
waterfall_plot(rc, rc_date=RCDT, rc_sum=RCTLSUM, rc_resp=RCRESP)
rc \%>\% 
  left_join(arm, by="SUBJID") \%>\%
  waterfall_plot(rc_date=RCDT, rc_sum=RCTLSUM, rc_resp=RCRESP, arm=ARM) +
  NULL
}
