% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_grade.R
\name{ae_plot_grade}
\alias{ae_plot_grade}
\title{Graphic representation of AEs}
\usage{
ae_plot_grade(
  df_ae,
  ...,
  df_enrol,
  variant = c("max", "sup", "eq"),
  position = c("dodge", "stack", "fill"),
  type = c("relative", "absolute"),
  arm = NULL,
  grade = "AEGR",
  subjid = "SUBJID"
)
}
\arguments{
\item{df_ae}{adverse event dataset, one row per AE, containing subjid, soc, and grade}

\item{...}{unused}

\item{df_enrol}{enrollment dataset, one row per patient, containing subjid (and arm if needed). All patients should be in this dataset.}

\item{position}{Position adjustment (cf. \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}})}

\item{type}{whether to present patients as proportions (\code{relative}) or as counts (\code{absolute})}

\item{arm}{name of the treatment column in \code{df_enrol}. Case-insensitive. Can be set to \code{NULL} to not group.}

\item{grade}{name of the AE grade column in \code{df_ae}. Case-insensitive.}

\item{subjid}{name of the patient ID in both \code{df_ae} and \code{df_enrol}. Case-insensitive.}
}
\value{
a ggplot
}
\description{
Produce a graphic representation of AE, counting AE as bars for each patient, colored by grade. Can be faceted by treatment arm.
}
\examples{
tm = edc_example_ae()
load_list(tm)
ae_plot_grade(df_ae=ae, df_enrol=enrolres)
ae_plot_grade(df_ae=ae, df_enrol=enrolres, arm="ARM", variant=c("sup", "max"))
ae_plot_grade(df_ae=ae, df_enrol=enrolres, arm="ARM", type="absolute")
ae_plot_grade(df_ae=ae, df_enrol=enrolres, arm="ARM", position="fill")
ae_plot_grade(df_ae=ae, df_enrol=enrolres, arm="ARM", position="stack", type="absolute")
}
