% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_soc.R
\name{ae_table_soc}
\alias{ae_table_soc}
\alias{as_flextable.ae_table_soc}
\title{Summary tables for AE by SOC}
\usage{
ae_table_soc(
  df_ae,
  ...,
  df_enrol,
  arm = "ARM",
  term = "AETERM",
  soc = "AESOC",
  grade = "AEGR",
  subjid = "SUBJID",
  sort_by_ae = TRUE,
  total = TRUE,
  digits = 0,
  warn_miss = FALSE
)

\method{as_flextable}{ae_table_soc}(x, arm_colors = c("#f2dcdb", "#dbe5f1", "#ebf1dd", "#e5e0ec"))
}
\arguments{
\item{df_ae}{adverse event dataset, one row per AE, containing subjid, soc, and grade}

\item{...}{unused}

\item{df_enrol}{enrollment dataset, one row per patient, containing subjid (and arm if needed). All patients should be in this dataset.}

\item{arm}{name of the treatment column in \code{df_enrol}. Case-insensitive. Can be set to \code{NULL} to not group.}

\item{term}{name of the the CTCAE term column in \code{df_ae}. Case-insensitive. Can be set to \code{NULL}.}

\item{soc}{name of the SOC column in \code{df_ae}. Case-insensitive. Grade will be considered 0 if missing(e.g. if patient if absent from \code{df_ae}).}

\item{grade}{name of the AE grade column in \code{df_ae}. Case-insensitive.}

\item{subjid}{name of the patient ID in both \code{df_ae} and \code{df_enrol}. Case-insensitive.}

\item{sort_by_ae}{should the table be sorted by number or alphabetically}

\item{total}{whether to add a \code{total} column for each arm}

\item{digits}{significant digits for percentages}

\item{warn_miss}{whether to warn for missing values}

\item{x}{a dataframe, resulting of \code{ae_table_soc()}}

\item{arm_colors}{colors for the arm groups}
}
\value{
a dataframe (\code{ae_table_soc()}) or a flextable (\code{as_flextable()}).

a formatted flextable
}
\description{
The function \code{ae_table_soc()} creates a summary table of maximum AE grades for each patient according to term and SOC CTCAE.
The resulting dataframe can be piped to \code{as_flextable()} to get a nicely formatted flextable.
}
\examples{

tm = edc_example_ae()
ae_table_soc(df_ae=tm$ae, df_enrol=tm$enrolres, term=NULL)
ae_table_soc(df_ae=tm$ae, df_enrol=tm$enrolres, term=NULL, arm=NULL)

\dontrun{
#the resulting flextable can be customized using the flextable package
library(flextable)
ae_table_soc(tm$ae, df_enrol=tm$enrolres, total=FALSE) \%>\% 
  as_flextable() \%>\% 
  hline(i=~soc=="" & soc!=dplyr::lead(soc))
ae_table_soc(tm$ae, df_enrol=tm$enrolres, term=NULL, sort_by_ae=FALSE) \%>\% 
  as_flextable() \%>\% 
  hline()
ae_table_soc(tm$ae, df_enrol=tm$enrolres, term=NULL, arm=NULL) \%>\% 
  as_flextable()
}
}
\seealso{
\code{\link[=ae_table_grade_max]{ae_table_grade_max()}}, \code{\link[=ae_table_grade_n]{ae_table_grade_n()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade_max]{ae_plot_grade_max()}}, \code{\link[=ae_plot_grade_n]{ae_plot_grade_n()}}
}
