% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_soc.R
\name{ae_table_soc}
\alias{ae_table_soc}
\alias{as_flextable.ae_table_soc}
\title{Summary tables for AE by SOC}
\usage{
ae_table_soc(
  df_ae,
  df_enrol,
  arm = "ARM",
  term = "AETERM",
  soc = "AESOC",
  grade = "AEGR",
  subjid = "SUBJID",
  sort_by_ae = TRUE,
  total = TRUE,
  digits = 0,
  warn_miss = FALSE
)

as_flextable.ae_table_soc(
  x,
  arm_colors = c("#f2dcdb", "#dbe5f1", "#ebf1dd", "#e5e0ec")
)
}
\arguments{
\item{df_ae}{adverse event table, one row per AE, containing subjid, soc, and grade}

\item{df_enrol}{enrollment table, one row per patient, containing subjid (and arm if needed)}

\item{arm, term}{name of the treatment and term columns in \code{df}. Case-insensitive. Can be set to \code{NULL}.}

\item{subjid, soc, grade}{names of the other relevant columns in \code{df}. Case-insensitive.}

\item{sort_by_ae}{should the table be sorted by number or alphabetically}

\item{total}{should there be a \code{total} column for each arm}

\item{digits}{significant digits for percentages}

\item{x}{a dataframe, resulting of \code{ae_table_soc()}}

\item{arm_colors}{colors for the arm groups}
}
\value{
a dataframe (\code{ae_table_soc()}) or a flextable (\code{as_flextable()}).

a formatted flextable
}
\description{
The function \code{ae_table_soc()} creates a summary table of maximum AE grades for each patient according to term and SOC CTCAE.
The resulting dataframe can be piped to \code{as_flextable()} to get a nicely formatted flextable.
}
\examples{
\dontrun{
# 1) Create data
#enrolres = enrollment table, with one row per patient
#ae = adverse event table, with one row per AE

# 2) Apply functions
#the resulting flextable can be customized using the flextable package
ae_table_soc(ae, enrolres, total=FALSE) \%>\% 
  as_flextable() \%>\% flextable::hline(i=~soc=="" & soc!=lead(soc))
ae_table_soc(ae, enrolres, term=NULL, sort_by_ae=FALSE) \%>\% 
  as_flextable() \%>\% flextable::hline()
ae_table_soc(ae, enrolres, arm=NULL) \%>\% 
  as_flextable()
}
}
