% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_example}
\alias{data_example}
\alias{edc_example_mixed}
\alias{edc_example_plot}
\title{Example databases}
\usage{
edc_example_mixed(N = 100)

edc_example_plot(N = 50, seed = 42)
}
\arguments{
\item{N}{the number of patients}

\item{seed}{the random seed}
}
\value{
a list of tables
}
\description{
List of tables used in EDCimport examples.
}
