% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_example}
\alias{data_example}
\alias{edc_example_mixed}
\alias{edc_example_plot}
\alias{edc_example}
\title{Example databases}
\usage{
edc_example_mixed(N = 100)

edc_example_plot(N = 50, seed = 42)

edc_example(N = 50, seed = 42)
}
\arguments{
\item{N}{the number of patients}

\item{seed}{the random seed}
}
\value{
a list of tables
}
\description{
List of tables used in EDCimport examples:
\itemize{
\item \code{edc_example()} can be used as the result of \code{\link[=read_trialmaster]{read_trialmaster()}}
\item \code{edc_example_plot()} can be used to test \code{\link[=edc_swimmerplot]{edc_swimmerplot()}}
\item \code{edc_example_mixed()} can be used to test \code{\link[=split_mixed_datasets]{split_mixed_datasets()}}
}
}
