% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{edc_data_warn}
\alias{edc_data_warn}
\title{Standardized warning system}
\usage{
edc_data_warn(
  df,
  message,
  ...,
  issue_n = NULL,
  max_subjid = 5,
  col_subjid = get_subjid_cols()
)
}
\arguments{
\item{df}{the filtered dataframe}

\item{message}{the message. Can use {cli} formats.}

\item{...}{unused}

\item{issue_n}{(optional) identifying row number}

\item{max_subjid}{max number of subject ID to show in the message}

\item{col_subjid}{column name for subject ID}
}
\value{
nothing
}
\description{
Database issues should be traced in a separate table file, with an identifying row number
}
\examples{
tm = edc_example_mixed()
a = tm$long_pure \%>\% dplyr::filter(val1a>2)
edc_data_warn(a, "{.val val1} should be lesser than 2", issue_n=1)
}
