% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{edc_lookup}
\alias{edc_lookup}
\alias{get_lookup}
\title{Retrieve the lookup table from options}
\usage{
edc_lookup(..., check_null = TRUE)
}
\arguments{
\item{...}{passed on to \code{\link[dplyr:arrange]{dplyr::arrange()}}}

\item{check_null}{whether to stop if lookup is NULL}
}
\value{
the lookup dataframe summarizing the database import
}
\description{
Retrieve the lookup table from options
}
\examples{
tm = edc_example()
load_list(tm)
edc_lookup()
edc_lookup(dataset)
}
\seealso{
\code{\link[=build_lookup]{build_lookup()}}, \code{\link[=extend_lookup]{extend_lookup()}}
}
