% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{edc_viewer}
\alias{edc_viewer}
\title{Shiny data explorer}
\usage{
edc_viewer(background = TRUE)
}
\arguments{
\item{background}{should the app run in a background process}
}
\description{
Run a Shiny application that allows to browse the datasets.
}
