% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDCimport-package.R
\docType{package}
\name{EDCimport-package}
\alias{EDCimport}
\alias{EDCimport-package}
\title{EDCimport: Import Data from EDC Software}
\description{
A convenient toolbox to import data exported from Electronic Data Capture (EDC) software 'TrialMaster' and 'Macro'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DanChaltiel/EDCimport}
  \item Report bugs at \url{https://github.com/DanChaltiel/EDCimport/issues}
}

}
\author{
\strong{Maintainer}: Dan Chaltiel \email{dan.chaltiel@gmail.com} (\href{https://orcid.org/0000-0003-3488-779X}{ORCID})

}
\keyword{internal}
