% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDCimport-package.R
\docType{package}
\name{EDCimport-package}
\alias{EDCimport}
\alias{EDCimport-package}
\title{EDCimport: Import data from EDC software}
\description{
EDCimport is a package designed to easily import data from EDC software TrialMaster and Macro.
}
\author{
\strong{Maintainer}: Dan Chaltiel \email{dan.chaltiel@gmail.com} (\href{https://orcid.org/0000-0003-3488-779X}{ORCID})

}
\keyword{internal}
