% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extend_lookup}
\alias{extend_lookup}
\title{Extend the lookup table}
\usage{
extend_lookup(
  lookup,
  id = getOption("edc_id", "SUBJID"),
  datasets = get_datasets(),
  crfname = getOption("edc_crfname", "crfname")
)
}
\arguments{
\item{lookup}{}

\item{id}{}
}
\value{
the lookup, extended
}
\description{
This utility extands the lookup table to include the name
}
