% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_keyword}
\alias{find_keyword}
\title{Find a keyword in names or labels of
Trouver un mot-clé dans les noms ou labels des tables chargées}
\usage{
find_keyword(keyword, data = getOption("tablelist", NULL), ignore_case = TRUE)
}
\arguments{
\item{keyword}{the keyword to search for}

\item{data}{the dataframe where to look, generated by \code{\link[=read_trialmaster]{read_trialmaster()}}}

\item{ignore_case}{should case differences be ignored in the match?}
}
\value{
a filtered tibble
}
\description{
Find a keyword in names or labels of
Trouver un mot-clé dans les noms ou labels des tables chargées
}
\examples{
library(crosstable)
library(dplyr)
library(purrr)
#for this example, tablelist comes from `{crosstable}` but it should come 
#from [read_trialmaster()] and be left as default
mtcars2 = crosstable::mtcars2
iris2 = crosstable::iris2
tablelist = tibble(
  dataset=c("mtcars2", "iris2"),
  names=map(dataset, ~names(get(.x))), 
  labels=map(dataset, ~labelled::var_label(get(.x)))
)
find_keyword("hp", data=tablelist)
find_keyword("number|date", data=tablelist)
find_keyword("number|date", data=tablelist, ignore_case=FALSE)
}
