% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_key_cols}
\alias{get_key_cols}
\title{Important column names}
\usage{
get_key_cols(
  patient_id = getOption("edc_id", c("ptno", "subjid")),
  crfname = getOption("edc_crfname", "crfname"),
  ...
)
}
\arguments{
\item{patient_id}{the name of the columns containing the patient ID}

\item{crfname}{the name of the columns containing the name of the CRF page}

\item{...}{unused}
}
\description{
Important column names
}
