% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_lookup}
\alias{get_lookup}
\title{Generate a lookup table}
\usage{
get_lookup(data_list)
}
\arguments{
\item{data_list}{a list containing at least 1 dataframe}
}
\value{
a dataframe summarizing column names and labels
}
\description{
Generate a lookup table
}
\examples{
x = edc_example()
x$.lookup=NULL
lk = get_lookup(x)
lk
lk \%>\% tidyr::unnest(c(names, labels))

}
