% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_lookup}
\alias{get_lookup}
\title{Generate a lookup table}
\usage{
get_lookup(data_list)
}
\arguments{
\item{data_list}{a list containing at least 1 dataframe}
}
\value{
a dataframe summarizing column names and labels
}
\description{
Generate a lookup table
}
\examples{
tl = list(r=crosstable::iris2, x=mtcars) \%>\% get_lookup()
tl
library(tidyr)
tl \%>\% unnest(everything()) \%>\% unnest(everything())
}
