% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_lookup}
\alias{get_lookup}
\title{Retrieve the lookup table from options}
\usage{
get_lookup()
}
\value{
the lookup dataframe summarizing column names and labels
}
\description{
Retrieve the lookup table from options
}
