% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_subjid_cols}
\alias{get_subjid_cols}
\title{Get key column names}
\usage{
get_subjid_cols(lookup = edc_lookup())
}
\arguments{
\item{lookup}{the lookup table}
}
\value{
a character vector
}
\description{
Retrieve names of patient ID and CRF name from the actual names of the datasets, without respect of the case. Default values should be set through options.
}
\section{options}{

Use \code{edc_options()} to set default values:
\itemize{
\item \code{edc_cols_subjid} defaults to \code{c("SUBJID", "PTNO")}
\item \code{edc_cols_crfname} defaults to \code{c("FORMDESC", "CRFNAME")}
}
}

\examples{
#get_subjid_cols()
#get_crfname_cols()
}
\keyword{internal}
