% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_subjid_cols}
\alias{get_subjid_cols}
\alias{get_crfname_cols}
\title{Get key column names}
\usage{
get_subjid_cols(
  subjid_cols = getOption("edc_cols_subjid", c("PTNO", "SUBJID")),
  lookup = get_lookup()
)

get_crfname_cols(
  crfname_cols = getOption("edc_cols_crfname", "CRFNAME"),
  lookup = get_lookup()
)
}
\arguments{
\item{subjid_cols, crfname_cols}{the possible column names that can hold key data. Case-insensitive.}

\item{lookup}{the lookup table}
}
\value{
a character vector
}
\description{
Retrieve names of patient ID (usually "SUBJID" and "PATNO") and CRF name (usually "CRFNAME") from the actual names of the datasets, without respect of the case. Default values can (and should) be set through options.
}
\examples{
get_subjid_cols()
get_crfname_cols()
}
