% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_newer_data.R
\name{search_for_newer_data}
\alias{search_for_newer_data}
\title{Search for newer data}
\usage{
search_for_newer_data(
  project_name,
  path = c("data", path_home("Downloads")),
  ask = TRUE,
  advice = NULL
)
}
\arguments{
\item{project_name}{the project name, treated as a regex}

\item{path}{the path vector to be searched, default to both "data" and the usual "Downloads" folder}

\item{ask}{whether to ask the user to move the file to "data"}

\item{advice}{whether to advice how to move it, if \code{ask==FALSE}}
}
\value{
the path to the newer file, invisibly.
}
\description{
Search in some folders if a TrialMaster database more recent than the current extration is present. By default, it will search the "data" folder and the OS usual "Downloads" folder. If a newer database is found, user will be asked if they want to move it to the "data" folder.
}
