% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_distinct}
\alias{select_distinct}
\title{Select only distinct columns}
\usage{
select_distinct(df, .by)
}
\arguments{
\item{df}{a dataframe}

\item{.by}{optional grouping columns}
}
\value{
\code{df} with less columns
}
\description{
Select all columns that has only one level for a given grouping scope.
Useful when dealing with mixed datasets containing both long data and repeated short data.
}
\examples{
tm = edc_example_ae()
tm$ae \%>\% names
tm$ae \%>\% select_distinct() \%>\% names
tm$ae \%>\% select_distinct(.by=subjid) \%>\% names
}
