% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_mixed.R
\name{split_mixed_datasets}
\alias{split_mixed_datasets}
\title{Split mixed datasets}
\usage{
split_mixed_datasets(
  id,
  ...,
  verbose = TRUE,
  output_code = FALSE,
  lookup = getOption("edc_lookup", NULL)
)
}
\arguments{
\item{id}{the patient identifier, probably "SUBJID". Should be shared by all datasets.}

\item{...}{not used}

\item{verbose}{whether to print informations about the process.}

\item{output_code}{whether to print the code to explictly write. Can also be a file path.}

\item{lookup}{the lookup table}
}
\value{
a list of the new long and short tables. Use \code{\link[=load_list]{load_list()}} to load them in the global environment.
}
\description{
Split mixed tables, i.e. tables that hold both long data (N values per patient) and short data (one value per patient, duplicated on N lines), into one long table and one short table.
}
\examples{
#tm = read_trialmaster("filename.zip", pw="xx")
tm = edc_example_mixed()
names(tm)
load_list(tm)
print(long_mixed) #`val1` and `val2` are long but `val3` is short

mixed_data = split_mixed_datasets("SUBJID", lookup=.lookup)
load_list(mixed_data)
print(long_mixed_short) 
print(long_mixed_long) 

#alternatively, get the code and only use the one you need
split_mixed_datasets("SUBJID", lookup=.lookup, output_code=TRUE)
split_mixed_datasets("SUBJID", lookup=.lookup, output_code="mixed_code.R")#' 
}
