% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{swimmerplot}
\alias{swimmerplot}
\title{Get a swimmer plot of all dates in the database}
\usage{
swimmerplot(
  .lookup = getOption("edc_lookup", NULL),
  id = "SUBJID",
  plotly = TRUE
)
}
\arguments{
\item{.lookup}{the result of \code{\link[=get_lookup]{get_lookup()}}}

\item{id}{an identifier for a given row}

\item{plotly}{whether to use {plotly} to get an interactive plot}
}
\value{
a plot
}
\description{
Get a swimmer plot of all dates in the database
}
