% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{swimmerplot}
\alias{swimmerplot}
\title{Get a swimmer plot of all dates in the database}
\usage{
swimmerplot(
  .lookup = getOption("edc_lookup", NULL),
  id = "SUBJID",
  origin = NULL,
  aes_color = c("variable", "label"),
  plotly = TRUE
)
}
\arguments{
\item{.lookup}{the lookup table, loaded along with the database or result of \code{\link[=get_lookup]{get_lookup()}}}

\item{id}{an identifier for a given row}

\item{origin}{a variable to consider as the time 0, given as "dataset$column"}

\item{aes_color}{either \code{variable} ("{dataset} - {column}") or \code{label} (column label)}

\item{plotly}{whether to use \code{{plotly}} to get an interactive plot}
}
\value{
a plot
}
\description{
Get a swimmer plot of all dates in the database
}
