% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{waterfall_plot}
\alias{waterfall_plot}
\title{Generate a waterfall plot}
\usage{
waterfall_plot(
  data_recist,
  rc_sum = "RCTLSUM",
  rc_resp = "RCRESP",
  rc_date = "RCDT",
  type = c("best_resp", "worst_resp"),
  rc_star = NULL,
  arm = NULL,
  warn_missing = TRUE
)
}
\arguments{
\item{data_recist}{recist dataset}

\item{rc_sum}{name of the target lesions length sum column in \code{data_recist}, usually "RCTLSUM".}

\item{rc_resp}{name of the response column in \code{data_recist}, usually "RCRESP".}

\item{rc_date}{name of the date column in \code{data_recist}, usually "RCDT".}

\item{type}{one of \code{c("best_resp", "worst_resp")}}

\item{rc_star}{name of the logical column in \code{data_recist} that triggers the .  Can be set to \code{NULL}.}

\item{arm}{name of the treatment column in \code{data_recist}. Can be set to \code{NULL} to not group.}

\item{warn_missing}{whether to warn about missing values}
}
\value{
a ggplot
}
\description{
Generate a waterfall plot
}
\section{Methods}{

Data are ordered on \code{rc_date}.
}

\examples{

\dontrun{
waterfall_plot(rc, rc_date="RCDT", rc_sum="RCTLSUM", rc_resp="RCRESP")
waterfall_plot(rc, rc_date="RCDT", rc_sum="RCTLSUM", rc_resp="RCRESP", type="worst_resp")
rc \%>\% 
  left_join(arm, by="SUBJID") \%>\%
  mutate(star = any(RCNEW=="1-Yes", na.rm=TRUE), .by=SUBJID) \%>\% 
  waterfall_plot(rc_date="RCDT", rc_sum="RCTLSUM", rc_resp="RCRESP", arm="ARM", rc_star="star")
}
}
